完善ok

+++ resources/views/user/index.blade.php
@@ -32,6 +32,7 @@
             <th class="sorting" role="columnheader" tabindex="0" aria-controls="DataTables_Table_0" rowspan="1" colspan="1" aria-label="Browser: activate to sort column ascending" style="width: 224px;">用户名</th>
             <th class="sorting" role="columnheader" tabindex="0" aria-controls="DataTables_Table_0" rowspan="1" colspan="1" aria-label="Platform(s): activate to sort column ascending" style="width: 209px;">邮箱</th>
             <th class="sorting" role="columnheader" tabindex="0" aria-controls="DataTables_Table_0" rowspan="1" colspan="1" aria-label="Engine version: activate to sort column ascending" style="width: 148px;">状态</th>
+            <th class="sorting" role="columnheader" tabindex="0" aria-controls="DataTables_Table_0" rowspan="1" colspan="1" aria-label="Engine version: activate to sort column ascending" style="width: 148px;">角色</th>
             <th class="sorting" role="columnheader" tabindex="0" aria-controls="DataTables_Table_0" rowspan="1" colspan="1" aria-label="CSS grade: activate to sort column ascending" style="width: 111px;">操作</th></tr>
         </thead>
         <tbody role="alert" aria-live="polite" aria-relevant="all">
@@ -45,11 +46,18 @@
 	            <td class=" sorting_1">{{$v['id']}}</td>
 	            <td class=" ">{{$v['username']}}</td>
 	            <td class=" ">{{$v['email']}}</td>
-	            <td class=" ">{{showstatus($v['status'])}}</td>
+              <td class=" ">{{showstatus($v['status'])}}</td>
+              @if($v['display_name']==NULL)
+              <td class=" ">没有角色</td>
+              @else
+              <td class=" " style="color:red"><a href="/admins/user/roledel?uid={{$v['id']}}&role_id={{$v['role_id']}}" onclick="return confirm('确定要删除该角色吗?')">{{$v['display_name']}}</a></td>
+              @endif
 	            <td class=" ">
 	            <!-- 修改和删除的连接 -->
-	            <a href="/admins/user/edit/{{$v['id']}}" style='color:black;font-size:20px'><i class="icon-pencil-2"></i></a>&nbsp;&nbsp;&nbsp;&nbsp;
-	            <a href="/admins/user/delete/{{$v['id']}}" style='color:black;font-size:20px'><i class="icon-trash"></i></a>
+	            <a href="/admins/user/edit?id={{$v['id']}}&rol={{$v['display_name']}}&role_id={{$v['role_id']}}" style='color:black;font-size:20px'><i class="icon-pencil-2"></i></a>&nbsp;&nbsp;&nbsp;&nbsp;
+	            <a href="/admins/user/delete?id={{$v['id']}}" style='color:black;font-size:20px' onclick="return confirm('确定删除该用户和该用户下的所有角色吗?')"><i class="icon-trash"></i></a>&nbsp;&nbsp;&nbsp;&nbsp;
+              <a href="/Role/attach?id={{$v['id']}}" style='color:black;font-size:20px'><i class="icon-user"></i></a>
+
 	            </td>
 	          </tr>
 	        @endforeach
